Update ts utils and fix apps (#216)
* update ts-utils (update array utils)

* fix apps to follow ts-utils array updates

* update ts-utils (use int type)

* fix apps to follow ts-utils int type updates

+++ packages/apps/catan_dice_app/src/functions/roll-dice.ts
@@ -1,6 +1,6 @@
-import { randInt } from '@noshiro/ts-utils';
+import { randInt, uint32 } from '@noshiro/ts-utils';
 
 export const rollTwoDices = (): [number, number] => [
-  randInt(1, 6),
-  randInt(1, 6),
+  randInt(1 as uint32, 6 as uint32),
+  randInt(1 as uint32, 6 as uint32),
 ];
