+++ packages/utils/ts-utils/src/array/mutation/remove-at.test.ts
@@ -1,8 +1,9 @@
+import { uint32 } from '../../types';
 import { removeAt } from './remove-at';
 
 test('removeAt middle', () => {
   const array = [1, 2, 3, 4, 5];
-  const index = 2;
+  const index = 2 as uint32;
   const result = removeAt(array, index);
   expect(array).toEqual([1, 2, 4, 5]);
   expect(result).toBe(3);
@@ -10,7 +11,7 @@ test('removeAt middle', () => {
 
 test('removeAt first', () => {
   const array = [1, 2, 3, 4, 5];
-  const index = 0;
+  const index = 0 as uint32;
   const result = removeAt(array, index);
   expect(array).toEqual([2, 3, 4, 5]);
   expect(result).toBe(1);
@@ -18,7 +19,7 @@ test('removeAt first', () => {
 
 test('removeAt last', () => {
   const array = [1, 2, 3, 4, 5];
-  const index = 4;
+  const index = 4 as uint32;
   const result = removeAt(array, index);
   expect(array).toEqual([1, 2, 3, 4]);
   expect(result).toBe(5);
@@ -26,7 +27,7 @@ test('removeAt last', () => {
 
 test('removeAt out of range', () => {
   const array = [1, 2, 3, 4, 5];
-  const index = 9;
+  const index = 9 as uint32;
   const result = removeAt(array, index);
   expect(array).toEqual([1, 2, 3, 4, 5]);
   expect(result).toBe(undefined);
@@ -34,7 +35,7 @@ test('removeAt out of range', () => {
 
 test('removeAt out of range', () => {
   const array: number[] = [];
-  const index = 1;
+  const index = 1 as uint32;
   const result = removeAt(array, index);
   expect(array).toEqual([]);
   expect(result).toBe(undefined);
