+++ packages/utils/ts-utils/src/array/slice/skip.test.ts
@@ -1,32 +1,33 @@
+import { uint32 } from '../../types';
 import { testArrayEquality } from '../array-tester';
 import { skip } from './skip';
 
 testArrayEquality({
   testName: 'skipLast',
-  target: skip([1, 2, 3, 4, 5], 3),
+  target: skip([1, 2, 3, 4, 5], 3 as uint32),
   toBe: [4, 5],
 });
 
 testArrayEquality({
   testName: 'skip',
-  target: skip([1, 2, 3, 4, 5], 0),
+  target: skip([1, 2, 3, 4, 5], 0 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
 
 testArrayEquality({
   testName: 'skip',
-  target: skip([1, 2, 3, 4, 5], -1),
+  target: skip([1, 2, 3, 4, 5], -1 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
 
 testArrayEquality({
   testName: 'skip',
-  target: skip([1, 2, 3, 4, 5], 5),
+  target: skip([1, 2, 3, 4, 5], 5 as uint32),
   toBe: [],
 });
 
 testArrayEquality({
   testName: 'skip',
-  target: skip([1, 2, 3, 4, 5], 6),
+  target: skip([1, 2, 3, 4, 5], 6 as uint32),
   toBe: [],
 });
