+++ packages/utils/ts-utils/src/date/functions/get-all-dates-of-month.ts
@@ -1,13 +1,15 @@
 import { range } from '../../array';
+import { uint32 } from '../../types';
 import { DateEnum, MonthEnum, YearEnum } from '../types';
 import { getLastDateNumberOfMonth } from './get-last-date-number-of-month';
 import { newDate } from './new-date';
 import { toTimestamp } from './to-timestamp';
 
 export const getAllDatesOfMonth = (year: YearEnum, month: MonthEnum): Date[] =>
-  range(1, getLastDateNumberOfMonth(year, month) + 1).map((date) =>
-    newDate(year, month, date as DateEnum)
-  );
+  range(
+    1 as uint32,
+    (getLastDateNumberOfMonth(year, month) + 1) as uint32
+  ).map((date) => newDate(year, month, date as DateEnum));
 
 export const getAllDatesOfMonthAsTimestamp = (
   year: YearEnum,
